// src/app/interfaces/AppRoutes.ts
export enum AppRoutes {
  HOME = '',
  WELCOME = 'welcome',
  SEARCH = 'search',
  CREATE_EVENT = 'create-event',
  CREATE_EVENT_STEP1 = 'create-event/step1',
  CREATE_EVENT_STEP2 = 'create-event/step2',
  CREATE_EVENT_STEP3 = 'create-event/step3',
  MY_EVENTS = 'my-events',
  MY_EVENTS_REQUESTS = 'my-events/my-requests',
  MY_SPACE = 'my-space',
  EVENT = 'event/:id',
  EVENT_CHAT = 'event/:id/chat',
  EVENT_LISTS = 'event/:id/lists',
  EVENT_LISTS_OVERVIEW = 'event/:id/lists/',
  EVENT_LISTS_DETAIL = 'event/:id/lists/detail/:listId',
  EVENT_SURVEYS = 'event/:id/surveys',
  EVENT_REQUESTS = 'event/:id/requests',
  EVENT_GUESTS = 'event/:id/guests',
  PROFILE = 'profile/:id',
  STYLING = 'styling',
  LEGAL_DISCLOSURE = 'legal-disclosure',
  PRIVACY_POLICY = 'privacy-policy',
  TERMS_AND_CONDITIONS = 'terms-and-conditions',
  NOT_FOUND = '404',
  SETTINGS = 'settings',
  INFORMATION = 'information',
  ACCOUNT = 'account',
  SHARE_PROFILE = 'share-profile',
  ADD_FRIEND = 'add-friend/:username/:inviteId',
  WILDCARD = '**',
}